Refactor. Placing classes in separate modules, e.g. Glacier and SurgingGlacier in glacier, GlacierBed in glacierBed, MassBalance in massBalance and GlacierCollection in glacierCollection.
