+++ oggm_edu/__init__.py
@@ -1,4 +1,7 @@
 __version__ = '0.0.1'
 
 from .funcs import *
-from .glacier import *
+from oggm_edu.glacier import Glacier, SurgingGlacier
+from oggm_edu.glacierBed import GlacierBed
+from oggm_edu.glacierCollection import GlacierCollection
+from oggm_edu.massBalance import MassBalance
