+++ oggm_edu/glacier.py
@@ -1,7 +1,15 @@
-'''This class provides a high level interface to the OGGM modelling framwork,
+'''This module provides a high level interface to the OGGM modelling framwork,
 suitable for educational purposes regarding the understanding the basics of
 glacier mechanics.
+
+It contains two classes: Glacier and SurgingGlacier, the top-level interfaces
+to experimenting with glaciers through OGGM-Edu. Initialisation depends on
+two other classes: the GlacierBed and the MassBalance.
 '''
+
+# Internals
+from oggm_edu.glacierBed import GlacierBed
+from oggm_edu.massBalance import MassBalance
 
 # Other libraries.
 import numpy as np
@@ -9,7 +17,6 @@ import seaborn as sns
 import xarray as xr
 import pandas as pd
 import warnings
-import collections
 import copy
 
 # Plotting
@@ -17,7 +24,6 @@ from matplotlib import pyplot as plt
 from matplotlib.patches import Patch
 # Import OGGM things.
 from oggm.core.flowline import FluxBasedModel, RectangularBedFlowline
-from oggm.core.massbalance import LinearMassBalance
 
 from oggm import cfg
 # Initialise cfg
@@ -25,280 +31,6 @@ cfg.initialize_minimal()
 sns.set_context('notebook')
 sns.set_style('ticks')
 plt.rcParams['figure.figsize'] = (12, 9)
-
-
-class GlacierBed:
-    '''The glacier bed'''
-
-    def __init__(self, top=None, bottom=None, width=None, altitudes=None,
-                 widths=None, slope=None, nx=200, map_dx=100):
-        '''Initialise the bed. Pass single scalars for top, bottom and width
-         to create a square glacier bed. For finer control pass altitudes and
-         widths as lists or tuples for custom geometry. Will linearly
-         intepolate between the altitude/width pairs.
-
-        Parameters
-        ----------
-        top : int or float
-            Elevation at the top of the bed domain in meters.
-        bottom : int or float
-            Elevation at the bottom of the bed domain in meters.
-        width : int or float
-            Width of the bed in meters. Generates a square bed.
-        altitudes : array_like (ints or floats)
-            list of values corresponding to the altitude/width distribution.
-            First and last value will be the top and bottom.
-            Length should match widths.
-        widths : array_like, (ints of floats)
-            List of values defining the widths along the glacier.
-            Length should match altitudes.
-        slope : float
-            Define the slope of the bed, decimal
-        nx : int
-            Number of grid points.
-        map_dx : int
-            Grid point spacing in meters.
-        '''
-        # Arg checks
-        if (top is None and altitudes is None) or \
-                (bottom is None and altitudes is None) or \
-                (top is not None and altitudes is not None) or \
-                (bottom is not None and altitudes is not None):
-            raise ValueError('Provide either a top and bottom or altitudes.')
-
-        if (width is None and widths is None) or \
-                (width is not None and widths is not None):
-            raise ValueError('Provide either a single width'
-                             ' or a list of widths')
-        if (altitudes is not None and width is not None):
-            raise ValueError('Not possible to combine altitude with width.')
-        if (top is not None and bottom is not None and widths is not None):
-            raise ValueError('Not possible to combine widths '
-                             'with top and bottom.')
-
-        # Check that width is a single scalar.
-        if width is not None:
-            # Ask for forgiveness...
-            try:
-                width + 1
-            except TypeError:
-                raise TypeError('Width should be single scalar (int/float)')
-
-        # Geometry
-        # If we have top and bottom.
-        if top and bottom:
-            self.top = top
-            self.bottom = bottom
-        # Else get it from the altitudes.
-        else:
-            self.top = altitudes[0]
-            self.bottom = altitudes[-1]
-
-        # Check the values of top and bottom.
-        if self.top <= self.bottom and self.bottom < 0:
-            raise ValueError('Top of the bed has to be above the bottom.' +
-                             ' Bottom also has to be above 0')
-        # Calculate the slope correction
-        if slope:
-            if slope >= 1. or slope <= 0.:
-                raise ValueError('Slope shoud be between 0 and 1 (not equal)')
-            # If there is a slop we re calculate the map_dx
-            else:
-                map_dx = (self.top - self.bottom) / (nx * slope)
-
-        # Set the resolution.
-        self.map_dx = map_dx
-        # Initialise the bed
-        self.bed_h = np.linspace(self.top, self.bottom, nx)
-        self.distance_along_glacier = np.linspace(0, nx, nx) *\
-            self.map_dx * 1e-3
-        # Widths.
-        # If width has a length of one, we have a constant width.
-        if width:
-            self.width = width
-            self.widths = (np.zeros(nx) + self.width) / map_dx
-        # If length of width and length of width altitudes are compatible,
-        # we create a bed with variable width.
-        elif len(altitudes) == len(widths):
-            self.width = widths
-            # First create a constant bed.
-            tmp_w = np.zeros(nx)
-
-            # Make sure we have lists.
-            altitudes = list(altitudes)
-            widths = list(widths)
-            # Check that the provided altitudes make sense.
-            altitudes_tmp = altitudes.copy()
-            altitudes_tmp.sort(reverse=True)
-            if not altitudes_tmp == altitudes:
-                raise ValueError('Please provides altitudes'
-                                 ' in descending order.')
-            # Loop over the altitude/width pairs and do a linear
-            # interpolations between them.
-            for i, alt in enumerate(altitudes[1:]):
-                # We want to interpolate to the previos altitude step.
-                inter_top = altitudes[i]
-                # Select the values that we interpolate.
-                mask = np.logical_and(self.bed_h <= inter_top,
-                                      self.bed_h >= alt)
-                # Linear interpolation between the widths.
-                tmp_w[mask] = np.linspace(self.width[i], self.width[i+1],
-                                          mask.sum())
-            # Assign the varied widths it.
-            self.widths = tmp_w / map_dx
-
-        # If noting works, raise an error.
-        else:
-            raise ValueError('Provided arguments are not compatible.')
-
-    def __repr__(self):
-
-        # Get the json representation of the object.
-        json = self._to_json()
-
-        # Create a nicer string of it.
-        string = 'Glacier bed \n'
-        for key, value in json.items():
-            string += f'{key}: {value} \n'
-        return string
-
-    def _repr_html_(self):
-        '''HTML representations'''
-        # Get attris
-        attrs = self._to_json()
-        df = pd.DataFrame.from_dict(attrs, orient="index")
-        df.columns = [""]
-        df.index.name = type(self).__name__
-
-        return df._repr_html_()
-
-    def _to_json(self):
-        '''Json representation of the bed'''
-        # Is the bed width constant or variable?
-        w_string = ('constant' if type(self.width) in (int, float)
-                    else 'variable')
-        # Create a dictionary with some of the attributes.
-        json = {
-            'Bed type': f'Linear bed with a {w_string} width',
-            'Top [m]': self.top,
-            'Bottom [m]': self.bottom,
-            'Width(s) [m]': [self.width],
-            'Length [km]': self.distance_along_glacier[-1]
-        }
-        return json
-
-    def plot(self):
-        '''Plot the glacier bed'''
-
-        fig, (ax1, ax2) = plt.subplots(nrows=2,
-                                       gridspec_kw={'height_ratios': [2, 1]},
-                                       sharex=True)
-
-        # Plot the bed
-        ax1.plot(self.distance_along_glacier, self.bed_h, label='Bedrock',
-                 ls=':', c='k', lw=2, zorder=3)
-        # And fill it.
-        ax1.fill_betweenx(self.bed_h, self.distance_along_glacier,
-                          color='lightgrey')
-        # Some labels etc.
-        ax1.set_ylabel('Altitude [m]')
-        ax1.set_facecolor('#ADD8E6')
-        ax1.set_ylim(self.bottom, self.top + 400)
-
-        # Fill the bed.
-        ax2.fill_between(self.distance_along_glacier,
-                         -self.widths/2 * self.map_dx,
-                         self.widths/2 * self.map_dx,
-                         color='lightgrey')
-        # More styling.
-        ax2.set_facecolor('darkgrey')
-        ax2.axhline(0, c='k')
-        plt.xlim((0, self.distance_along_glacier[-1] + 2))
-        plt.xlabel('Distance along glacier [km]')
-        plt.ylabel('Distance from centerline [m]')
-        ax1.set_title('Glacier domain')
-        ax1.legend()
-
-
-class MassBalance(LinearMassBalance):
-    '''Mass balance class'''
-
-    def __init__(self, ELA, gradient, hemisphere=None,
-                 density=None):
-        '''
-        Initialise the mass balance from the ELA and the gradient.
-        '''
-        super().__init__(ela_h=ELA, grad=gradient)
-
-        # Temperature bias evolution
-        self._temp_bias_final = 0.
-        self._temp_bias_grad = 0.
-        self._temp_bias_intersect = 0.
-        self._temp_bias_final_year = 0
-
-    def _to_json(self):
-        json = {
-            'ELA [m]': self.ela_h,
-            'Original ELA [m]': self.orig_ela_h,
-            'Temperature bias [C]': self.temp_bias,
-            'Gradient [mm/m/yr]': self.grad,
-            'Hemisphere': self.hemisphere,
-            'Ice density [kg/m3]': self.rho
-        }
-        return json
-
-    def __repr__(self):
-        json = self._to_json()
-        string = 'Glacier mass balance \n'
-        # Create a nice string
-        for key, value in json.items():
-            string += f'{key}: {value} \n'
-        return string
-
-    def _repr_html_(self):
-        '''HTML representations'''
-        # Get attris
-        attrs = self._to_json()
-        df = pd.DataFrame.from_dict(attrs, orient="index")
-        df.columns = [""]
-        df.index.name = type(self).__name__
-
-        return df._repr_html_()
-
-    def _update_temp_bias(self, year):
-        '''Updates the temperature bias internally during progression.'''
-        # Change the temperature bias
-        new_temp_bias = self._temp_bias_grad * year + self._temp_bias_intersect
-        self.temp_bias = new_temp_bias
-
-    def _add_temp_bias(self, temp_bias, duration, year):
-        '''Method to add a gradual temperature bias to the mass balance of the
-        glacier.
-
-        Parameters
-        ----------
-        temp_bias: float
-            Final temperature bias to apply after the specified duration
-        duration: int
-            Number of year during which to apply the temperature bias.
-        year: int
-            Current age of the glacier
-        '''
-        # Check that criteria are met.
-        if isinstance(temp_bias, float) and isinstance(duration, int):
-            # Set the temperature bias.
-            self._temp_bias_final = temp_bias
-            # Set the final year
-            self._temp_bias_final_year = year + duration
-            # Calculate the gradient, linear for now.
-            self._temp_bias_grad = (self._temp_bias_final - self.temp_bias)\
-                / duration
-            # And the intersect
-            self._temp_bias_intersect = self._temp_bias_final -\
-                (self._temp_bias_grad * self._temp_bias_final_year)
-
-        else:
-            raise TypeError('Temperature bias and/or duration of wrong type')
 
 
 class Glacier:
@@ -1121,250 +853,3 @@ class SurgingGlacier(Glacier):
                    bbox_to_anchor=(0.9, 0.89))
 
         plt.show()
-
-
-class GlacierCollection:
-    '''This is class for storing multiple glaciers, providing convenient
-    methods for comparing them.
-    '''
-
-    def __init__(self, glacier_list=None):
-        '''Initialise the glacier collection.
-
-        Parameters
-        ----------
-        glacier_list: list of glaciers objects
-            Defaults to none. Has the possibility to add glaciers on the go.
-        '''
-
-        self._glaciers = []
-
-        # Do we have a list of glaciers?
-        if glacier_list:
-            # Check that all are Glaciers.
-            if all(isinstance(glacier, Glacier) for glacier in glacier_list):
-                self._glaciers = glacier_list
-
-    def __repr__(self):
-        return f'Glacier collection with {len(self._glaciers)} glaciers.'
-
-    def _repr_html_(self):
-        # Pretty representation for notebooks.
-        # Get the first glacier and base the creation of the dataframe of this.
-        # If we have glaciers in the collection.
-        if len(self._glaciers) > 0:
-            json = {**self._glaciers[0]._to_json(),
-                    **self._glaciers[0].bed._to_json(),
-                    **self._glaciers[0].mass_balance._to_json()
-                    }
-            # Create the dataframe.
-            df = pd.DataFrame(json, index=[0])
-            # We then loop over the rest of the collection.
-            for glacier in self._glaciers[1:]:
-                # Next json representation
-                json = {**glacier._to_json(), **glacier.bed._to_json(),
-                        **glacier.mass_balance._to_json()}
-                # Add it to the df.
-                df = df.append(json, ignore_index=True)
-                df.index.name = "Glacier"
-            # Return the html representation of the dataframe.
-            return df._repr_html_()
-        # If empty.
-        else:
-            pass
-
-    def check_collection(self, glacier):
-        '''Utility method. Check if the glaciers obey the collection rules.
-        They need to have the same domains for this to make sense I think.'''
-        # TODO
-
-    @property
-    def glaciers(self):
-        return self._glaciers
-
-    def add(self, glacier):
-        '''Add a glacier to the collection.
-
-        Parameters
-        ----------
-        glacier: oggm.edu Glacier object or array_like with
-        Glacier objects
-        '''
-        # Check if iterable
-        if isinstance(glacier, collections.Sequence):
-            for glacier in glacier:
-                # Check that glacier is of the right type.
-                if not isinstance(glacier, Glacier):
-                    raise TypeError('Glacier collection can only'
-                                    ' contain glacier objects.')
-                # Is the glacier already in the collection?
-                if glacier in self._glaciers:
-                    raise AttributeError('Glacier is already in collection')
-                # If the glacier is an instance of Glacier, we can add it to
-                # the collection.
-                else:
-                    self._glaciers.append(glacier)
-        # If not iterable
-        else:
-            # Check that glacier is of the right type.
-            if not isinstance(glacier, Glacier):
-                raise TypeError('Glacier collection can only'
-                                ' contain glacier objects.')
-            # Is the glacier already in the collection?
-            if glacier in self._glaciers:
-                raise AttributeError('Glacier is already in collection')
-            # If the glacier is an instance of Glacier, we can add it to
-            # the collection.
-            else:
-                self._glaciers.append(glacier)
-
-    def progress_to_year(self, year):
-        '''Progress the glaciers within the collection to
-        the specified year.
-
-        Parameters
-        ----------
-        year: int
-            Which year to grow the glaciers.
-        '''
-        if len(self._glaciers) < 1:
-            raise ValueError('Collection is empty')
-
-        # Loop over the glaciers within the collection.
-        for glacier in self._glaciers:
-            glacier.progress_to_year(year)
-
-    def progress_to_equilibrium(self):
-        '''Progress the glaciers within the collection to equilibrium.
-        '''
-        if len(self._glaciers) < 1:
-            raise ValueError('Collection is empty')
-
-        # Loop.
-        for glacier in self._glaciers:
-            glacier.progress_to_equilibrium()
-
-    def plot(self):
-        '''Plot the glaciers in the collection to compare them.'''
-
-        if len(self._glaciers) < 1:
-            raise ValueError('Collection is empty')
-        # We use this to plot the bedrock etc.
-        gl1 = self._glaciers[0]
-        # Get the ax from the first plot
-        fig, ax = plt.subplots()
-        # Bedrock
-        ax.plot(gl1.bed.distance_along_glacier, gl1.bed.bed_h, label='Bedrock',
-                ls=':', c='k', lw=2, zorder=3)
-        ax.set_ylim((gl1.bed.bottom, gl1.bed.top + 200))
-        # Fill it in.
-        ax.fill_betweenx(gl1.bed.bed_h, gl1.bed.distance_along_glacier,
-                         facecolor='lightgrey')
-
-        # Set the title.
-        ax.set_title('Glacier collection')
-
-        elas = []
-        # Loop over the collection.
-        for i, glacier in enumerate(self._glaciers):
-            # Plot the surface
-            if glacier.current_state is not None:
-                # Masking shenanigans.
-                diff = glacier.current_state.surface_h - glacier.bed.bed_h
-                mask = diff > 0
-                idx = diff.argmin()
-                mask[:idx+1] = True
-                # Fill the ice.
-                ax.fill_between(glacier.bed.distance_along_glacier,
-                                glacier.current_state.surface_h,
-                                glacier.bed.bed_h,
-                                where=mask,
-                                facecolors='white',
-                                # edgecolors=color_cycler(i),
-                                # lw=2,
-                                )
-                # Plot outline
-                ax.plot(glacier.bed.distance_along_glacier[mask],
-                        glacier.current_state.surface_h[mask],
-                        label=f'Glacier nr. {i+1} at year'
-                        + f' {glacier.age}')
-                # Ylim
-                ax.set_ylim((gl1.bed.bottom,
-                             gl1.current_state.surface_h[0]+200))
-            elas.append(glacier.ELA)
-
-        # Loop the unique ELAs.
-        for i, ela in enumerate(set(elas)):
-            # Plot the ELA
-            ax.axhline(ela, ls='--', zorder=1)
-            # Label if elas are equal.
-            if len(set(elas)) == 1:
-                ax.text(glacier.bed.distance_along_glacier[-1], ela + 10,
-                        'ELAs are equal', ha='right', va='bottom')
-            # If we have multiple ELAs.
-            else:
-                ax.text(glacier.bed.distance_along_glacier[-1], ela + 10,
-                        f'ELA  nr {i+1}', ha='right', va='bottom')
-
-        # axis labels.
-        ax.set_xlabel('Distance along glacer [km]')
-        ax.set_ylabel('Altitude [m]')
-        ax.set_xlim((0, gl1.bed.distance_along_glacier[-1]+2))
-        ax.set_facecolor('#ADD8E6')
-        plt.legend(loc='lower left')
-        # Add a second legend with infos.
-        # It would be cool to only show attributes that are different.
-        labels = []
-        for glacier in self._glaciers:
-            # Create the label
-            label = f'ELA: {glacier.ELA} \n' \
-                f'MB grad: {glacier.mb_gradient} \n' \
-                f'Age: {glacier.age} \n' \
-                f'Creep: {glacier.creep:.2e} \n' \
-                f'Sliding: {glacier.basal_sliding}'
-            # Append the label to the list.
-            labels.append(label)
-        # Get the handles back
-        handles, _ = ax.get_legend_handles_labels()
-        # Create the legend
-        fig.legend(handles[1:], labels, title='Glacier info',
-                   loc='upper left', bbox_to_anchor=(0.9, 0.89))
-
-    def plot_history(self):
-        '''Plot the histories of the collection'''
-
-        fig, (ax1, ax2, ax3) = plt.subplots(nrows=3, sharex=True)
-
-        for glacier in self._glaciers:
-            # Plot the length.
-            glacier.history.length_m.plot(ax=ax1)
-            # Plot the volume
-            glacier.history.volume_m3.plot(ax=ax2)
-            # Plot the area
-            glacier.history.area_m2.plot(ax=ax3,
-                                         label=f'ELA: {glacier.ELA} \n' +
-                                         f'MB grad: {glacier.mb_gradient} \n' +
-                                         f'Age: {glacier.age} \n' +
-                                         f'Creep: {glacier.creep:.2e} \n' +
-                                         f'Sliding: {glacier.basal_sliding}')
-        # Decorations
-        ax1.set_xlabel('')
-        ax1.set_title('Glacier collection evolution')
-        ax1.annotate('Glacier length', (0.98, 0.1), xycoords='axes fraction',
-                     bbox={'boxstyle': 'Round', 'color': 'lightgrey'},
-                     ha='right')
-        ax1.set_ylabel('Length [m]')
-        ax2.set_xlabel('')
-        ax2.annotate('Glacier volume', (0.98, 0.1), xycoords='axes fraction',
-                     bbox={'boxstyle': 'Round', 'color': 'lightgrey'},
-                     ha='right')
-        ax2.set_ylabel('Volume [m$^3$]')
-        ax3.set_xlabel('Year')
-        ax3.annotate('Glacier area', (0.98, 0.1), xycoords='axes fraction',
-                     bbox={'boxstyle': 'Round', 'color': 'lightgrey'},
-                     ha='right')
-        ax3.set_ylabel('Area [m$^2$]')
-        handels, labels = ax3.get_legend_handles_labels()
-        fig.legend(handels, labels, loc='upper left', ncol=1,
-                   title='Glacier info',
-                   bbox_to_anchor=(0.9, 0.89))
